Add `ZooSimulationService` for simulation logic and enhance repository and services for animal and food management
- Introduced `ZooSimulationService` to handle simulation logic for hunger, feeding, reproduction, and deaths.
- Updated interfaces and implementations for `IZooRepository`, `IAnimalService`, and `IFoodService` to support simulation needs.
- Refactored `ZooSimulationWorker` to delegate simulation tasks to `ZooSimulationService`.
- Adjusted food consumption logic to use `double` precision for consistency.
- Added methods for adding and removing animals in `AnimalService` and updated animal DTOs with IDs.

+++ ZooWebApi/Domain/Animal.cs
@@ -9,5 +9,5 @@ public abstract class Animal
     public string Species { get; set; }
     public AnimalType Type { get; set; }
     public int Hunger { get; set; } = default;
-    public abstract int Consume(int amount);
+    public abstract double Consume(double amount);
 }
