+++ ZooWebApi/Domain/Carnivore.cs
@@ -2,7 +2,7 @@ namespace ZooWebApi.Domain;
 
 public class Carnivore : Animal
 {
-    public override int Consume(int amount)
+    public override double Consume(double amount)
     {
         return amount * 3;
     }
