+++ ZooWebApi/Persistence/DatabaseZooRepository.cs
@@ -6,6 +6,6 @@ public class DatabaseZooRepository : IZooRepository
 {
     //TODO
     public List<Animal> Animals { get; set; }
-    public int FoodStock { get; set; }
+    public double FoodStock { get; set; }
     public int StandardFoodAmount { get; }
 }
