+++ ZooWebApi/Persistence/IZooRepository.cs
@@ -5,6 +5,6 @@ namespace ZooWebApi.Persistence;
 public interface IZooRepository
 {
     public List<Animal> Animals { get; set; }
-    public int FoodStock { get; set; }
+    public double FoodStock { get; set; }
     public int StandardFoodAmount { get; }
 }
