+++ ZooWebApi/Services/Contracts/IFoodService.cs
@@ -2,7 +2,7 @@ namespace ZooWebApi.Services.Contracts;
 
 public interface IFoodService
 {
-    public int GeFoodStock(); // in kg
+    public double GeFoodStock(); // in kg
     public void AddFood(int amount); // in kg
     public bool FeedAnimals(); // returns true if food was successfully consumed
 }
